Applied some bug fixes people have been discussing in Issues to keep the repo usable

+++ maskrcnn_benchmark/data/datasets/evaluation/coco/coco_eval.py
@@ -133,7 +133,7 @@ def prepare_for_coco_segmentation(predictions, dataset):
         # rles = prediction.get_field('mask')
 
         rles = [
-            mask_util.encode(np.array(mask[0, :, :, np.newaxis], order="F"))[0]
+            mask_util.encode(np.array(mask[0, :, :, np.newaxis], dtype=np.uint8, order="F"))[0]
             for mask in masks
         ]
         for rle in rles:
